Updates and improvements
- Fix some more missing hotkeys in Xfce Wayland (such as brightness
keys), as well as adding options for customising mouse speed,
acceleration and scroll direction, as Xfce's own graphical mouse
settings are X11-only.

- Fix PipeWire ALSA compatibility by symlinking some needed
configuration files.

- Bump Shim from 15.8 to 16.1 now that a GRUB instability issue with
Shim 16.1 has been fixed. This Shim build (from Fedora) is also now
ual-signed by the Microsoft 2011 and 2023 secure boot certs, allowing
it to be booted on systems with only the latter installed. This update
also (finally) removes our reliance on the GRUB peimage module.

- Add zram-generator for convenience (disabled by default but has a
sample config file installed which can be uncommented to enable some
safe defaults).

- Ensure the build systems of some packages do not hardcode the running
kernel version and/or hostname of the system used for building.

- Re-add OpenSSL 3 libcrypto and libssl libraries (only) for binary
compatibility purposes. They will be removed once a newer release of
VMware is updated to be linked to OpenSSL 4.

+++ create-livecd.sh
@@ -184,7 +184,14 @@ else
   chmod +x iso-workdir/iso-root/EFI/BOOT/{BOOTAA64.EFI,{grubaa64,mmaa64}.efi}
 fi
 sed -e "s|@@VERSION@@|$ver|g" -e "s|@@ISOFILE@@|$isoname|g" livecd-data/grub.cfg.in > iso-workdir/iso-root/grub.cfg
-cp iso-workdir/massos-rootfs/usr/share/licenses/shim/copyright iso-workdir/iso-root/LICENSES/shim.txt
+if test -f iso-workdir/massos-rootfs/usr/share/licenses/shim/copyright; then
+  cp iso-workdir/massos-rootfs/usr/share/licenses/shim/copyright iso-workdir/iso-root/LICENSES/shim.txt
+elif test -f iso-workdir/massos-rootfs/usr/share/licenses/shim/COPYRIGHT; then
+  cp iso-workdir/massos-rootfs/usr/share/licenses/shim/COPYRIGHT iso-workdir/iso-root/LICENSES/shim.txt
+else
+  echo "Error: No license file found for shim." >&2
+  exit 1
+fi
 cp iso-workdir/massos-rootfs/usr/share/licenses/grub/COPYING iso-workdir/iso-root/LICENSES/GRUB.txt
 # Install Memtest86+, IPXE and UEFI EDK2 Shell.
 # Some of these won't be installed depending on architecture (if unsupported).
